refactor writeTestDataToDisk to reuse generate.Writer
